Try to support Froyo Apps2SD.
We listened the ACTION_EXTERNAL_APPLICATIONS_AVAILABLE and ACTION_EXTERNAL_APPLICATIONS_UNAVAILABLE intents,
so we should be able to load/unload those packages on the model.
Take care, AOSP 2.2 Launcher uses a totally revamped Model with differen threads and a lot of things like
making "disabled" icons for the missing apps and so on... i'm not sure if my patch is the only thing needed.
Widgets installed into SD are not reloaded/unloaded by this....