Patch to allow  DeleteZone resize the current desktop instead of showing up over it.
Cause of previously moving the Deletezone over the top of the screen we created a little problem:
sometimes it's hard to move items to the top desktop row.

So this patch just resizes the whole desktop screen when the DeleteZone shows.
Please test it and watch for possible caveats before merging it...